Solve git problem
깃이 꼬이면서 발생한 문제해결
