+++ Everyone-Tayo-Application/app/build.gradle
@@ -34,6 +34,9 @@ android {
     buildFeatures {
         viewBinding true
     }
+    viewBinding {
+        enabled = true
+    }
 }
 
 dependencies {
