+++ Everyone-Tayo-Application/app/src/main/AndroidManifest.xml
@@ -9,30 +9,19 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Everyonetayo">
-        <activity android:name=".KeyPad"></activity>
-        <activity android:name=".BusDriver" />
-        <activity
-<<<<<<< HEAD
-            android:name=".ui.login.LoginActivity"
+        <activity android:name=".view.KeyPad"></activity>
+        <activity android:name=".view.BusDriver" />
+        <activity android:name=".ui.login.LoginActivity"
             android:label="@string/title_activity_login" />
-        <activity android:name=".SignUp" />
-        <activity android:name=".SelectService" />
-        <activity android:name=".ChoiceBus" />
-        <activity android:name=".DropOff" />
-        <activity android:name=".VoiceCheck" />
-        <activity android:name=".VoiceReader" />
-        <activity android:name=".WaitingTime" />
-        <activity android:name=".MainActivity">
-=======
-
-            android:name=".view.login.LoginActivity"
-            android:label="@string/title_activity_login"></activity>
         <activity android:name=".view.SignUp" />
         <activity android:name=".view.SelectService" />
         <activity android:name=".view.ChoiceBus" />
+        <activity android:name=".view.DropOff" />
+        <activity android:name=".view.VoiceCheck" />
+        <activity android:name=".view.VoiceReader" />
+        <activity android:name=".view.WaitingTime" />
         <activity android:name=".view.MainActivity">
 
->>>>>>> b0fba00d85b310a03218410ef52525ac854d3496
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
