--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/BusDriver.kt
@@ -1,11 +1,0 @@
-package team.sw.everyonetayo
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-
-class BusDriver : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_bus_driver)
-    }
-}
