--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/DropOff.kt
@@ -1,40 +1,0 @@
-package team.sw.everyonetayo
-
-import android.content.Intent
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-import android.view.MenuItem
-import kotlinx.android.synthetic.main.activity_selectservice.*
-
-class DropOff : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_dropoff)
-
-        setTitle("하차 서비스")
-
-
-        home.setOnClickListener{
-            val intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-
-        //뒤로가기 버튼
-        supportActionBar?.setDisplayHomeAsUpEnabled(true);
-    }
-
-
-    //뒤로가기 버튼 동작 코드
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item?.itemId){
-            android.R.id.home -> {
-                finish()
-                return true
-            }
-            else -> {
-                return super.onOptionsItemSelected(item)
-            }
-        }
-    }
-}
