--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/SignUp.kt
@@ -1,31 +1,0 @@
-package team.sw.everyonetayo
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-import android.view.MenuItem
-
-class SignUp : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.signup)
-
-        setTitle("회원가입")
-
-        //뒤로가기 버튼
-        supportActionBar?.setDisplayHomeAsUpEnabled(true);
-    }
-
-
-    //뒤로가기 버튼 동작 코드
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item?.itemId){
-            android.R.id.home -> {
-                finish()
-                return true
-            }
-            else -> {
-                return super.onOptionsItemSelected(item)
-            }
-        }
-    }
-}
