--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/ui/login/LoggedInUserView.kt
@@ -1,9 +1,0 @@
-package team.sw.everyonetayo.ui.login
-
-/**
- * User details post authentication that is exposed to the UI
- */
-data class LoggedInUserView(
-    val displayName: String
-    //... other data fields that may be accessible to the UI
-)
