--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/ui/login/LoginFormState.kt
@@ -1,10 +1,0 @@
-package team.sw.everyonetayo.ui.login
-
-/**
- * Data validation state of the login form.
- */
-data class LoginFormState(
-    val usernameError: Int? = null,
-    val passwordError: Int? = null,
-    val isDataValid: Boolean = false
-)
