--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/ui/login/LoginResult.kt
@@ -1,9 +1,0 @@
-package team.sw.everyonetayo.ui.login
-
-/**
- * Authentication result : success (user details) or error message.
- */
-data class LoginResult(
-    val success: LoggedInUserView? = null,
-    val error: Int? = null
-)
