--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/ui/login/LoginViewModelFactory.kt
@@ -1,25 +1,0 @@
-package team.sw.everyonetayo.ui.login
-
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
-import team.sw.everyonetayo.data.LoginDataSource
-import team.sw.everyonetayo.data.LoginRepository
-
-/**
- * ViewModel provider factory to instantiate LoginViewModel.
- * Required given LoginViewModel has a non-empty constructor
- */
-class LoginViewModelFactory : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(LoginViewModel::class.java)) {
-            return LoginViewModel(
-                loginRepository = LoginRepository(
-                    dataSource = LoginDataSource()
-                )
-            ) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
