+++ Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/view/BusDriver.kt
@@ -1,11 +1,12 @@
-package team.sw.everyonetayo
+package team.sw.everyonetayo.view
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import team.sw.everyonetayo.R
 
-class KeyPad : AppCompatActivity() {
+class BusDriver : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_key_pad)
+        setContentView(R.layout.activity_bus_driver)
     }
 }
