+++ Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/view/KeyPad.kt
@@ -4,9 +4,9 @@ import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import team.sw.everyonetayo.R
 
-class voice_reader : AppCompatActivity() {
+class KeyPad : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_voice_check)
+        setContentView(R.layout.activity_key_pad)
     }
 }
