+++ Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/view/MainActivity.kt
@@ -3,6 +3,7 @@ package team.sw.everyonetayo.view
 import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import android.support.v7.app.ActionBar
 import kotlinx.android.synthetic.main.activity_main.*
 import team.sw.everyonetayo.R
 import team.sw.everyonetayo.view.login.LoginActivity
@@ -21,5 +22,10 @@ class MainActivity : AppCompatActivity() {
             val intent = Intent(this, LoginActivity::class.java)
             startActivity(intent)
         }
+
+        var actionBar : ActionBar?
+        actionBar = supportActionBar
+        actionBar?.hide()
+
     }
 }
