+++ Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/view/VoiceCheck.kt
@@ -1,18 +1,18 @@
-package team.sw.everyonetayo
+package team.sw.everyonetayo.view
 
 import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
-import kotlinx.android.synthetic.main.activity_choicebus.*
-import kotlinx.android.synthetic.main.activity_selectservice.home
+import kotlinx.android.synthetic.main.activity_select_service.*
+import team.sw.everyonetayo.R
 
-class ChoiceBus : AppCompatActivity() {
+class VoiceCheck : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_choicebus)
+        setContentView(R.layout.activity_voice_check)
 
-        setTitle("버스 종류 선택")
+        setTitle("음성인식 번호 확인")
 
 
         home.setOnClickListener{
@@ -21,28 +21,9 @@ class ChoiceBus : AppCompatActivity() {
             finish()
         }
 
-
-        regular_bus.setOnClickListener{
-            val intent = Intent(this, VoiceReader::class.java)
-            startActivity(intent)
-        }
-
-
-        village_bus.setOnClickListener{
-            val intent = Intent(this, VoiceReader::class.java)
-            startActivity(intent)
-        }
-
-
-        express_bus.setOnClickListener{
-            val intent = Intent(this, VoiceReader::class.java)
-            startActivity(intent)
-        }
-
         //뒤로가기 버튼
         supportActionBar?.setDisplayHomeAsUpEnabled(true);
     }
-
 
     //뒤로가기 버튼 동작 코드
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
