--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/view/voice_check.kt
@@ -1,12 +1,0 @@
-package team.sw.everyonetayo.view
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-import team.sw.everyonetayo.R
-
-class voice_check : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_voice_reader)
-    }
-}
