--- Everyone-Tayo-Application/app/src/main/java/team/sw/everyonetayo/view/waiting_time.kt
@@ -1,12 +1,0 @@
-package team.sw.everyonetayo.view
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-import team.sw.everyonetayo.R
-
-class waiting_time : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_waiting_time)
-    }
-}
