+++ Everyone-Tayo-Application/app/src/main/res/layout/activity_bus_driver.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BusDriver">
+    tools:context=".view.BusDriver">
 
 </android.support.constraint.ConstraintLayout>
