+++ Everyone-Tayo-Application/app/src/main/res/layout/activity_key_pad.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#4699FF"
-    tools:context=".KeyPad">
+    tools:context=".view.KeyPad">
 
     <TextView
         android:id="@+id/textView12"
