+++ Everyone-Tayo-Application/app/src/main/res/layout/activity_select_service.xml
@@ -5,29 +5,23 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#4699FF"
-
     tools:context=".view.SelectService">
 
-
-    <ImageView
-        android:id="@+id/imageView10"
-        android:layout_width="411dp"
-        android:layout_height="598dp"
-        app:srcCompat="@android:drawable/screen_background_light"
-        tools:layout_editor_absoluteX="0dp"
-        tools:layout_editor_absoluteY="67dp" />
-
     <Button
-        android:id="@+id/button3"
-        android:layout_width="412dp"
-        android:layout_height="140dp"
-        android:layout_marginTop="428dp"
+        android:id="@+id/stopover"
+        android:layout_width="319dp"
+        android:layout_height="121dp"
+        android:layout_marginStart="44dp"
+        android:layout_marginTop="440dp"
+        android:layout_marginEnd="48dp"
         android:background="@drawable/stopover"
-        android:fontFamily="@font/nanum_square_eb"
+        android:fontFamily="@font/nanum_square_r"
         android:text="하차\n서비스"
         android:textColor="#005490"
         android:textSize="40sp"
         android:textStyle="bold"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         tools:ignore="HardcodedText,MissingConstraints" />
@@ -36,37 +30,66 @@
         android:id="@+id/textView10"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/nanum_square_eb"
+        android:layout_marginStart="17dp"
+        android:layout_marginTop="60dp"
+        android:layout_marginEnd="22dp"
+        android:fontFamily="@font/nanum_square_r"
         android:text='"버스 하차를 원하시나요?"'
-        android:textColor="#005595"
+        android:textColor="#FFFFFF"
         android:textSize="34sp"
-        tools:layout_editor_absoluteX="19dp"
-        tools:layout_editor_absoluteY="382dp" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/riding2" />
+
+    <Button
+        android:id="@+id/riding2"
+        android:layout_width="319dp"
+        android:layout_height="121dp"
+        android:layout_marginStart="46dp"
+        android:layout_marginTop="200dp"
+        android:layout_marginEnd="46dp"
+        android:background="@drawable/riding"
+        android:fontFamily="@font/nanum_square_r"
+        android:text="승차\n서비스"
+        android:textColor="#FFFFFF"
+        android:textSize="40sp"
+        android:textStyle="bold"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:ignore="HardcodedText,MissingConstraints" />
+
+    <Button
+        android:id="@+id/home"
+        android:layout_width="100dp"
+        android:layout_height="100dp"
+        android:layout_marginStart="155dp"
+        android:layout_marginTop="68dp"
+        android:layout_marginEnd="156dp"
+        android:background="@drawable/home_button"
+        android:drawableTop="@drawable/home_size"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/stopover" />
 
     <TextView
         android:id="@+id/textView11"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/nanum_square_eb"
+        android:layout_marginStart="17dp"
+        android:layout_marginTop="152dp"
+        android:layout_marginEnd="22dp"
+        android:fontFamily="@font/nanum_square_r"
         android:text='"버스 승차를 원하시나요?"'
-        android:textColor="#005490"
-        android:textSize="34sp"
-        tools:layout_editor_absoluteX="19dp"
-        tools:layout_editor_absoluteY="148dp" />
-
-    <Button
-        android:id="@+id/riding2"
-        android:layout_width="410dp"
-        android:layout_height="141dp"
-        android:layout_marginTop="192dp"
-        android:background="@drawable/riding"
-        android:fontFamily="@font/nanum_square_eb"
-        android:text="승차\n서비스"
         android:textColor="#FFFFFF"
-        android:textSize="40sp"
-        android:textStyle="bold"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:ignore="HardcodedText,MissingConstraints"
-        tools:layout_editor_absoluteX="0dp" />
+        android:textSize="34sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
 
 </android.support.constraint.ConstraintLayout>
