+++ Everyone-Tayo-Application/app/src/main/res/layout/activity_voice_check.xml
@@ -5,65 +5,71 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#4699FF"
+    tools:context=".view.VoiceCheck">
 
-    tools:context=".view.voice_reader">
-
-
-    <ImageView
-        android:id="@+id/imageView4"
-        android:layout_width="75dp"
-        android:layout_height="48dp"
-        app:srcCompat="@drawable/tayo_logo1"
-        tools:layout_editor_absoluteX="16dp"
-        tools:layout_editor_absoluteY="10dp" />
-
-    <ImageView
-        android:id="@+id/imageView9"
-        android:layout_width="411dp"
-        android:layout_height="598dp"
-        app:srcCompat="@android:drawable/screen_background_light"
-        tools:layout_editor_absoluteX="0dp"
-        tools:layout_editor_absoluteY="67dp" />
 
     <Button
-        android:id="@+id/button9"
-        android:layout_width="410dp"
-        android:layout_height="163dp"
-        android:layout_marginTop="400dp"
-        android:background="@drawable/stopover"
-        android:fontFamily="@font/nanum_square_round_eb"
-        android:text="아니오"
-        android:textAlignment="center"
-        android:textColor="#00307C"
-        android:textSize="48sp"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:layout_editor_absoluteX="1dp" />
+        android:id="@+id/home"
+        android:layout_width="100dp"
+        android:layout_height="100dp"
+        android:layout_marginStart="155dp"
+        android:layout_marginTop="89dp"
+        android:layout_marginEnd="156dp"
+        android:background="@drawable/home_button"
+        android:drawableTop="@drawable/home_size"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/button9" />
 
     <Button
         android:id="@+id/button7"
-        android:layout_width="410dp"
-        android:layout_height="162dp"
-        android:layout_marginTop="220dp"
+        android:layout_width="320dp"
+        android:layout_height="120dp"
+        android:layout_marginStart="51dp"
+        android:layout_marginTop="268dp"
+        android:layout_marginEnd="40dp"
         android:background="@drawable/riding"
-        android:fontFamily="@font/nanum_square_round_eb"
+        android:fontFamily="@font/nanum_square_r"
         android:text="예"
         android:textAlignment="center"
         android:textColor="#FFFFFF"
         android:textSize="48sp"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:layout_editor_absoluteX="0dp" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/textView3"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="40dp"
-        android:layout_marginTop="152dp"
-        android:fontFamily="@font/nanum_square_round_eb"
+        android:layout_marginTop="188dp"
+        android:layout_marginEnd="44dp"
+        android:fontFamily="@font/nanum_square_r"
         android:text='"0번 버스가 맞습니까?"'
         android:textAlignment="center"
-        android:textColor="#005490"
+        android:textColor="#FFFFFF"
         android:textSize="34sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/button9"
+        android:layout_width="320dp"
+        android:layout_height="120dp"
+        android:layout_marginStart="51dp"
+        android:layout_marginTop="408dp"
+        android:layout_marginEnd="40dp"
+        android:background="@drawable/stopover"
+        android:fontFamily="@font/nanum_square_r"
+        android:text="아니오"
+        android:textAlignment="center"
+        android:textColor="#00307C"
+        android:textSize="48sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
