+++ Everyone-Tayo-Application/app/src/main/res/layout/activity_voice_reader.xml
@@ -5,29 +5,38 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#4699FF"
+    tools:context=".view.VoiceReader">
 
-    tools:context=".view.voice_check">
-
-
-    <ImageView
-        android:id="@+id/imageView5"
-        android:layout_width="411dp"
-        android:layout_height="598dp"
-        app:srcCompat="@android:drawable/screen_background_light"
-        tools:layout_editor_absoluteX="0dp"
-        tools:layout_editor_absoluteY="67dp" />
+    <Button
+        android:id="@+id/keypad_call"
+        android:layout_width="247dp"
+        android:layout_height="75dp"
+        android:layout_marginStart="82dp"
+        android:layout_marginTop="24dp"
+        android:layout_marginEnd="82dp"
+        android:background="@drawable/press_button1"
+        android:fontFamily="@font/nanum_square_r"
+        android:text="키패드 입력"
+        android:textColor="#031170"
+        android:textSize="30sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/textView9" />
 
     <TextView
         android:id="@+id/textView8"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="60dp"
-        android:layout_marginTop="132dp"
-        android:fontFamily="@font/nanum_square_round_eb"
+        android:layout_marginTop="96dp"
+        android:layout_marginEnd="65dp"
+        android:fontFamily="@font/nanum_square_r"
         android:text="버스의 번호를\n말해주세요."
         android:textAlignment="center"
-        android:textColor="#005490"
+        android:textColor="#FFFFFF"
         android:textSize="48sp"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -36,32 +45,29 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="96dp"
-        android:layout_marginTop="552dp"
-        android:fontFamily="@font/nanum_square_round_eb"
+        android:layout_marginTop="520dp"
+        android:layout_marginEnd="102dp"
+        android:fontFamily="@font/nanum_square_r"
         android:text="음성인식 중 . . ."
         android:textAlignment="center"
-        android:textColor="#005490"
+        android:textColor="#FFFFFF"
         android:textSize="34sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageView
-        android:id="@+id/imageView3"
-        android:layout_width="75dp"
-        android:layout_height="48dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="10dp"
+        android:id="@+id/imageView2"
+        android:layout_width="283dp"
+        android:layout_height="265dp"
+        android:layout_marginStart="63dp"
+        android:layout_marginTop="24dp"
+        android:layout_marginEnd="65dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/tayo_logo1" />
-
-    <ImageView
-        android:id="@+id/imageView12"
-        android:layout_width="314dp"
-        android:layout_height="257dp"
-        android:layout_marginStart="48dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/voice2"
-        tools:layout_editor_absoluteY="270dp" />
+        app:layout_constraintTop_toBottomOf="@+id/textView8"
+        app:srcCompat="@drawable/voice" />
 
 </android.support.constraint.ConstraintLayout>
