--- Everyone-Tayo-Application/app/src/main/res/layout/activity_voicecheck.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#4699FF"
-    tools:context=".VoiceCheck">
-
-
-    <Button
-        android:id="@+id/home"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:layout_marginStart="155dp"
-        android:layout_marginTop="89dp"
-        android:layout_marginEnd="156dp"
-        android:background="@drawable/home_button"
-        android:drawableTop="@drawable/home_size"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/button9" />
-
-    <Button
-        android:id="@+id/button7"
-        android:layout_width="320dp"
-        android:layout_height="120dp"
-        android:layout_marginStart="51dp"
-        android:layout_marginTop="268dp"
-        android:layout_marginEnd="40dp"
-        android:background="@drawable/riding"
-        android:fontFamily="@font/nanum_square_r"
-        android:text="예"
-        android:textAlignment="center"
-        android:textColor="#FFFFFF"
-        android:textSize="48sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/textView3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="40dp"
-        android:layout_marginTop="188dp"
-        android:layout_marginEnd="44dp"
-        android:fontFamily="@font/nanum_square_r"
-        android:text='"0번 버스가 맞습니까?"'
-        android:textAlignment="center"
-        android:textColor="#FFFFFF"
-        android:textSize="34sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/button9"
-        android:layout_width="320dp"
-        android:layout_height="120dp"
-        android:layout_marginStart="51dp"
-        android:layout_marginTop="408dp"
-        android:layout_marginEnd="40dp"
-        android:background="@drawable/stopover"
-        android:fontFamily="@font/nanum_square_r"
-        android:text="아니오"
-        android:textAlignment="center"
-        android:textColor="#00307C"
-        android:textSize="48sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-</android.support.constraint.ConstraintLayout>
